+++ src/migrations/2012_12_06_225988_migration_cartalyst_sentry_install_throttle.php
@@ -39,6 +39,7 @@ class MigrationCartalystSentryInstallThrottle extends Migration {
 			$table->boolean('banned')->default(0);
 			$table->timestamp('last_attempt_at')->nullable();
 			$table->timestamp('suspended_at')->nullable();
+			$table->timestamp('banned_at')->nullable();
 		});
 	}